backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pymupdf==1.23.7
python-docx==1.1.0
aiofiles==23.2.1
requests>=2.25.0
python-dotenv>=0.19.0